Refactor code for easier approach
