+++ submission_train_code.py
@@ -13,7 +13,11 @@ from utility.parser import Parser
 import coloredlogs
 coloredlogs.install(logging.DEBUG)
 
-# All the evaluations will be evaluated on MineRLObtainDiamond-v0 environment
+# --- NOTE ---
+# This code is only used for "Research" track submissions
+# ------------
+
+# All the evaluations will be evaluated on MineRLObtainDiamondVectorObf-v0 environment
 MINERL_GYM_ENV = os.getenv('MINERL_GYM_ENV', 'MineRLObtainDiamondVectorObf-v0')
 # You need to ensure that your submission is trained in under MINERL_TRAINING_MAX_STEPS steps
 MINERL_TRAINING_MAX_STEPS = int(os.getenv('MINERL_TRAINING_MAX_STEPS', 8000000))
@@ -22,7 +26,7 @@ MINERL_TRAINING_MAX_INSTANCES = int(os.getenv('MINERL_TRAINING_MAX_INSTANCES', 5
 # You need to ensure that your submission is trained within allowed training time.
 # Round 1: Training timeout is 15 minutes
 # Round 2: Training timeout is 4 days
-MINERL_TRAINING_TIMEOUT = int(os.getenv('MINERL_TRAINING_TIMEOUT_MINUTES', 4*24*60))
+MINERL_TRAINING_TIMEOUT = int(os.getenv('MINERL_TRAINING_TIMEOUT_MINUTES', 4 * 24 * 60))
 # The dataset is available in data/ directory from repository root.
 MINERL_DATA_ROOT = os.getenv('MINERL_DATA_ROOT', 'data/')
 
@@ -35,7 +39,7 @@ parser = Parser(
     maximum_steps=MINERL_TRAINING_MAX_STEPS,
     raise_on_error=False,
     no_entry_poll_timeout=600,
-    submission_timeout=MINERL_TRAINING_TIMEOUT*60,
+    submission_timeout=MINERL_TRAINING_TIMEOUT * 60,
     initial_poll_timeout=600
 )
 
@@ -75,7 +79,7 @@ def main():
     # Save trained model to train/ directory
     # Training 100% Completed
     aicrowd_helper.register_progress(1)
-    #env.close()
+    env.close()
 
 
 if __name__ == "__main__":
